backend/requirements.txt
asgiref==3.9.2
Django==4.2.24
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
Pillow==10.0.1
drf-spectacular==0.27.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
